Add generic post-processing routine
This adds the required code to plug in new post-processing routines on the
last acquired image. These post-processing routines could be used to apply
effects on the acquired video stream.

The selected post-processing filter is expected to save its output on the
matrix provided by the "cam" descriptor. This matrix is expected to be a
valid 3 channel RGB matrix, suitable to be composed on the display with
overlay information.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>