Bugfix out-of-range exception
Removing a set of transformations on video frames which randomly produces
an "out-of-range" exception during decoding. The nature of this exception is
not known, but this transformation also seem to be not needed and very
CPU intensive. Thus, this patch remove the exception by removing a piece
of examples copy-pasted code which however is not required. ;-)

Signed-off-by: Patrick Bellasi <derkling@gmail.com>