Scale down VIDEO frame on-demand
A new "reduce_fct" attribute is added to the camera structure. By setting
it to a value lower than 1.0, frames are scaled down just after acquisition,
thus _before_ any effect is applyied on the image.
This allows to have a set of "reduce resolutions" which are available
to trade resolution with effects given a predefined processing budget.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>