Add CAMERA resolution tuning
This add a specific support to tune the CAMERA resolution. Webcams usually
allows for an HW configuration of the frame size. For the time being, this
method exploit this support to reconfigure on-demand the webcam for the
required HIGH, MID or LOW resolution.

However, HW reconfiguration could be subject to high overheads, thus a
software downscaling (i.e. by properly setting "reduce_fct") could be
considered in case these latencies could be an issue.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>