Add CycleTime enforcing
If a certain frame-rate (FPS) is required, which by default is 25, this new
method allows to pad the cycle time spent on actual processing with a properly
dimensioned sleep time so that the actual frame-rate match the required one.

This padding is used right after a new image has been drawn on display, just
before to precede with the next frame.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>